feat(customer) : add api put customers with service unit-test

+++ src/main/java/ttb/assg/customer/model/entity/CustomerAddress.java
@@ -19,7 +19,7 @@ public class CustomerAddress implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "address_seq")
-    private Long addressSeq;
+    private Integer addressSeq;
 
     @Column(name = "address_type")
     private String addressType;
